+++ mindcode/src/main/java/info/teksol/mindcode/logic/MindustryOpcodeVariants.java
@@ -206,6 +206,40 @@ public class MindustryOpcodeVariants {
         add(list, V7, V7A, W, FUNC, Opcode.CUTSCENE,  cutscene("zoom"), in("level"));
         add(list, V7, V7A, W, FUNC, Opcode.CUTSCENE,  cutscene("stop"));
 
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("warn"),            in("x"), in("y"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("cross"),           in("x"), in("y"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("blockFall"),       in("x"), in("y"), unused("size"), unused("color"), in("blocktype"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("placeBlock"),      in("x"), in("y"), in("size"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("placeBlockSpark"), in("x"), in("y"), in("size"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("breakBlock"),      in("x"), in("y"), in("size"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("spawn"),           in("x"), in("y"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("trail"),           in("x"), in("y"), in("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("breakProp"),       in("x"), in("y"), in("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("smokeCloud"),      in("x"), in("y"), unused("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("vapor"),           in("x"), in("y"), unused("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("hit"),             in("x"), in("y"), unused("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("hitSquare"),       in("x"), in("y"), unused("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("shootSmall"),      in("x"), in("y"), in("rotation"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("shootBig"),        in("x"), in("y"), in("rotation"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("smokeSmall"),      in("x"), in("y"), unused("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("smokeBig"),        in("x"), in("y"), unused("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("smokeColor"),      in("x"), in("y"), in("rotation"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("smokeSquare"),     in("x"), in("y"), in("rotation"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("smokeSquareBig"),  in("x"), in("y"), in("rotation"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("spark"),           in("x"), in("y"), unused("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("sparkBig"),        in("x"), in("y"), unused("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("sparkShoot"),      in("x"), in("y"), in("rotation"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("sparkShootBig"),   in("x"), in("y"), in("rotation"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("drill"),           in("x"), in("y"), unused("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("drillBig"),        in("x"), in("y"), unused("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("lightBlock"),      in("x"), in("y"), in("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("explosion"),       in("x"), in("y"), in("size"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("smokePuff"),       in("x"), in("y"), unused("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("sparkExplosion"),  in("x"), in("y"), unused("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("crossExplosion"),  in("x"), in("y"), in("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("wave"),            in("x"), in("y"), in("size"), in("color"));
+        add(list, V7, V7A, W, FUNC, Opcode.EFFECT,  cutscene("bubble"),          in("x"), in("y"));
+
         add(list, V7, V7A, W, FUNC, Opcode.EXPLOSION, in("team"), in("x"), in("y"), in("radius"), in("damage"), in("air"), in("ground"), in("pierce"));
         add(list, V7, V7A, W, FUNC, Opcode.SETRATE, in("ipt"));
 
@@ -261,6 +295,10 @@ public class MindustryOpcodeVariants {
 
     public static NamedParameter draw(String name) {
         return new NamedParameter(LogicParameter.DRAW, name);
+    }
+
+    public static NamedParameter effect(String name) {
+        return new NamedParameter(LogicParameter.EFFECT, name);
     }
 
     public static NamedParameter fetch(String name) {